src/requirements.txt
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus
python-json-logger
fastapi
prometheus_client